Integrated Navigation Components and organized .kt files into appropriate packages
