+++ app/src/main/java/com/kafleyozone/coin/PreAuthenticationActivity.kt
@@ -1,27 +1,18 @@
 package com.kafleyozone.coin
 
-import android.os.Bundle
-import android.os.UserManager
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.add
-import androidx.fragment.app.commit
+import androidx.navigation.findNavController
+import com.kafleyozone.coin.fragments.ExitOnboardingDialogFragment
 
 class PreAuthenticationActivity : AppCompatActivity(R.layout.activity_pre_authentication) {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        if (savedInstanceState == null) {
-            supportFragmentManager.commit {
-                setReorderingAllowed(true)
-                add<OnboardingFlowFragment>(R.id.fragment_container_view, OnboardingFlowFragment.TAG)
-            }
-        }
-    }
 
     override fun onBackPressed() {
-        if (supportFragmentManager.findFragmentByTag(OnboardingFlowFragment.TAG)?.isVisible == true) {
-            ExitOnboardingDialogFragment().show(supportFragmentManager, ExitOnboardingDialogFragment.TAG)
-        } else {
-            super.onBackPressed()
+        when (findNavController(R.id.nav_host_fragment_container).currentDestination?.id) {
+            R.id.onboardingFlowFragment -> {
+                ExitOnboardingDialogFragment()
+                        .show(supportFragmentManager, ExitOnboardingDialogFragment.TAG)
+            } else -> super.onBackPressed()
         }
+
     }
 }
