+++ app/src/main/java/com/kafleyozone/coin/fragments/AccountSetupFragment.kt
@@ -1,4 +1,4 @@
-package com.kafleyozone.coin
+package com.kafleyozone.coin.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,7 +10,11 @@ import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.kafleyozone.coin.viewmodels.AccountSetupFragmentViewModel
+import com.kafleyozone.coin.models.BankInstitutionEntity
+import com.kafleyozone.coin.rvadapters.BankListAdapter
 import com.kafleyozone.coin.databinding.FragmentAccountSetupBinding
+import com.kafleyozone.coin.printListDebug
 
 class AccountSetupFragment(pagerListener: OnboardingFlowFragment.PagerListenerInterface) : Fragment() {
