+++ app/src/main/java/com/kafleyozone/coin/fragments/ExitOnboardingDialogFragment.kt
@@ -1,10 +1,11 @@
-package com.kafleyozone.coin
+package com.kafleyozone.coin.fragments
 
 import android.app.Dialog
 import android.content.DialogInterface
 import android.os.Bundle
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.DialogFragment
+import com.kafleyozone.coin.R
 import java.lang.IllegalStateException
 
 class ExitOnboardingDialogFragment : DialogFragment() {
