+++ app/src/main/java/com/kafleyozone/coin/fragments/HomeFragment.kt
@@ -1,0 +1,6 @@
+package com.kafleyozone.coin.fragments
+
+import androidx.fragment.app.Fragment
+import com.kafleyozone.coin.R
+
+class HomeFragment: Fragment(R.layout.fragment_home)
