+++ app/src/main/java/com/kafleyozone/coin/viewmodels/AccountSetupFragmentViewModel.kt
@@ -1,8 +1,9 @@
-package com.kafleyozone.coin
+package com.kafleyozone.coin.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.kafleyozone.coin.models.BankInstitutionEntity
 
 class AccountSetupFragmentViewModel : ViewModel() {
