+++ app/src/main/java/com/kafleyozone/coin/viewmodels/AddNewBankDialogFragmentViewModel.kt
@@ -1,10 +1,14 @@
-package com.kafleyozone.coin
+package com.kafleyozone.coin.viewmodels
 
 import android.util.Log
 import android.view.View
 import androidx.lifecycle.ViewModel
 import com.google.android.material.textfield.TextInputEditText
+import com.kafleyozone.coin.models.BankInstitutionEntity
+import com.kafleyozone.coin.R
 import com.kafleyozone.coin.databinding.FragmentAddNewBankDialogBinding
+import com.kafleyozone.coin.isMonetaryAmountValid
+import com.kafleyozone.coin.isNameValid
 import kotlin.NumberFormatException
 
 class AddNewBankDialogFragmentViewModel : ViewModel() {
