+++ app/src/main/res/layout/activity_pre_authentication.xml
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.fragment.app.FragmentContainerView
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:id="@+id/fragment_container_view" />
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/nav_host_fragment_container"
+    android:name="androidx.navigation.fragment.NavHostFragment"
+    tools:context=".PreAuthenticationActivity"
+    app:defaultNavHost="true"
+    app:navGraph="@navigation/nav_graph"/>
