+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    tools:context=".PreAuthenticationActivity"
+    app:startDestination="@id/onboardingFlowFragment">
+
+    <fragment
+        android:id="@+id/onboardingFlowFragment"
+        android:name="com.kafleyozone.coin.fragments.OnboardingFlowFragment"
+        android:label="OnboardingFlowFragment"
+        tools:layout="@layout/fragment_welcome">
+        <action
+            android:id="@+id/action_onboardingFlowFragment_to_loginFragment"
+            app:destination="@id/loginFragment"
+            app:popUpTo="@id/onboardingFlowFragment"
+            app:popUpToInclusive="false" />
+    </fragment>
+    <fragment
+        android:id="@+id/loginFragment"
+        android:name="com.kafleyozone.coin.fragments.LoginFragment"
+        android:label="LoginFragment"
+        tools:layout="@layout/fragment_login">
+        <action
+            android:id="@+id/action_loginFragment_to_homeFragment"
+            app:destination="@id/homeFragment"
+            app:popUpTo="@id/loginFragment"
+            app:popUpToInclusive="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="com.kafleyozone.coin.fragments.HomeFragment"
+        android:label="HomeFragment"
+        tools:layout="@layout/fragment_home"/>
+
+</navigation>
