+++ app/src/main/res/values/strings.xml
@@ -38,4 +38,5 @@
     <string name="add_bank_dialog_amount_tip">Specify an exact amount. For example: 562.36</string>
     <string name="input_validation_email_login">Enter the e-mail address associated with your Coin account.</string>
     <string name="input_validation_password_login">Enter your password.</string>
+    <string name="logout_button_label">Logout</string>
 </resources>
